+++ ptai-data-structures/src/test/java/com/ptsecurity/appsec/ai/ee/test/scan/brief/ScanBriefDetailedTest.java
@@ -35,13 +35,14 @@ public class ScanBriefDetailedTest extends BaseTest {
     }
 
     @Test
-    @DisplayName("Convert PT AI 4.1.1 and 4.2 scan results")
+    @DisplayName("Convert PT AI 4.1.1, 4.2.0 and 4.3.0 scan results")
     @SneakyThrows
     public void generateScanResults() {
         try (TempFile temp = TempFile.createFolder()) {
             Path briefDetailed = temp.toPath().resolve("brief").resolve("detailed");
             assertTrue(briefDetailed.toFile().mkdirs());
             for (ScanBrief.ApiVersion version : ScanBrief.ApiVersion.values()) {
+                if (version.isDeprecated()) continue;
                 Path destination = briefDetailed.resolve(version.name().toLowerCase());
                 assertTrue(destination.toFile().mkdirs());
                 for (Project project : ALL) {